+++ examples/tools/tool-package-quickstart/my_tool_package/tools/tool_with_custom_llm_type.py
